Extract signal queue handling into dedicated signals module
Moves signal queue init/teardown, push/drain/wait, and handle_pause
from main.py into signals.py. Reduces main.py from 1160 to 1104 lines
and creates a focused 83-line module with a single responsibility:
instant control signal delivery between HTTP endpoints and the agent loop.
